Remove old EventCard
